Bug fix
Fixed a bug where fever actions did not display correctly after previously failed fever
Fixed a bug where exit button did not work
Fixed a bug where fever progress was added twice

+++ Yakuza Dancing Game/Assets/Scripts/FeverController.cs
@@ -84,7 +84,7 @@ public class FeverController : MonoBehaviour
         OnFeverDisable.Invoke();
     }
 
-    public void AddFeverProgress(int amount)
+    public void AddFeverProgress(float amount)
     {
         _feverProgressBar.AddFill(amount);  // Add fill to fever progress bar
     }
